+++ tests/unit/search-service-tests.cpp
@@ -65,7 +65,8 @@ namespace
     [[nodiscard]] uburu::index::IndexUpdateSummary update(const uburu::WorktreeInfo&,
                                                           std::span<const uburu::FileEntry>,
                                                           const uburu::index::IndexProgressCallback& = {},
-                                                          std::stop_token = {}) override
+                                                          std::stop_token = {},
+                                                          const uburu::index::IndexUpdateOptions& = {}) override
     {
       return {};
     }
@@ -73,7 +74,8 @@ namespace
     [[nodiscard]] uburu::index::IndexUpdateSummary update(const uburu::WorktreeInfo&,
                                                           std::span<const uburu::index::IndexFileCandidate>,
                                                           const uburu::index::IndexProgressCallback& = {},
-                                                          std::stop_token = {}) override
+                                                          std::stop_token = {},
+                                                          const uburu::index::IndexUpdateOptions& = {}) override
     {
       return {};
     }
@@ -82,7 +84,8 @@ namespace
                                                           std::span<const uburu::FileEntry>,
                                                           std::span<const uburu::GitOverlayEntry>,
                                                           const uburu::index::IndexProgressCallback& = {},
-                                                          std::stop_token = {}) override
+                                                          std::stop_token = {},
+                                                          const uburu::index::IndexUpdateOptions& = {}) override
     {
       return {};
     }